requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Mail==0.9.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
Werkzeug==2.3.7
requests==2.31.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
gunicorn==21.2.0
